Fontelle on Windows: plugin editors, folders, dialogs, and a log to send
From the first person to run it on Windows: "this is what happens to
outside vsts in ur daw - it works but it's like only the knobs of like
every parameter ... the lfos say click to draw but seemingly does nothing
... anything that says click doesn't work ... oh now it crashed", and,
asked for the error log, "where get". Ty: "make the logs more accessible
for people maybe in a button on the home menu."

Plugin editors open on Windows. PluginWindow was X11-only, so every VST 3
and CLAP plugin fell back to the knob grid; it is now a Win32 window whose
HWND a VST 3 view, a CLAP win32 GUI or a bridged VST 2 editor embeds in,
driven by the studio's own message loop, and told the monitor's scale.

Windows has no HOME, and every folder Fontelle chose for itself came from
it: no settings file, presets, soundfont bank or crash reports. Settings
now live in %APPDATA%\fontelle and data in %LOCALAPPDATA%\fontelle.

The folder, open and save dialogs on Windows are the system's own
IFileDialog, owned by the window that asked, instead of PowerShell - which
froze the studio and put its dialog behind it. A settings button's answer
is a toast on the settings page, not a clipped status line.

Each windowed run tees stderr and stdout into logs/ under the data folder,
beside the crash reports, and the start menu has a Logs folder link. A
fault in native code (a plugin's DLL) now writes a report too, naming the
module on Windows, instead of reading as "ended from outside".

Pressing an LFO's picture starts drawing it, and a knob on a card nobody
can hear (level off, stage switched off) says so in its tip.

+++ crates/fontelle-host/src/bridge.rs
@@ -558,8 +558,9 @@ impl BridgedPlugin {
     }
 
     /// Opens the editor into `window`, answering the size the bridge says
-    /// it wants. The window's id is what an X11 embedding needs and what a
-    /// headless window reports as zero.
+    /// it wants. The window's id is what the platform embeds into — an X11
+    /// window, or on Windows an `HWND`, which is what `effEditOpen` takes —
+    /// and what a headless window reports as zero.
     pub(crate) fn open_editor(
         &self,
         window: &crate::gui::PluginWindow,
@@ -569,12 +570,7 @@ impl BridgedPlugin {
         }
         let (mut width, mut height) = (0u32, 0u32);
         let result = unsafe {
-            (self.table().open_editor)(
-                self.shared.instance,
-                u64::from(window.id()),
-                &mut width,
-                &mut height,
-            )
+            (self.table().open_editor)(self.shared.instance, window.id(), &mut width, &mut height)
         };
         if result < 0 {
             return Err(crate::gui::GuiError::Refused("create"));
